Title: 배열의 평균, Time: 15ms, Memory: 2MB, Status: Wrong Answer - CodeTree

+++ 241106/배열의 평균/ave-of-array.swift
@@ -2,17 +2,17 @@ import Foundation
 
 let n = 2
 
-var numbers = [[Int]]()
+var numbers = [[Double]]()
 
 for _ in 0 ..< n {
-    let input = readLine()!.trimmingCharacters(in: .whitespaces).components(separatedBy: " ").map { Int(String($0))! }
+    let input = readLine()!.trimmingCharacters(in: .whitespaces).components(separatedBy: " ").map { Double(String($0))! }
     numbers.append(input)
 }
 
 // 가로평균
 for number in numbers {
     let sum = number.reduce(0, +)
-    let average: Double = Double(sum / number.count)
+    let average: Double = sum / Double(number.count)
     print(average, terminator: " ")
 }
 
@@ -23,14 +23,14 @@ for index in 0 ..< 4 {
     let up = numbers[0][index]
     let down = numbers[1][index]
 
-    let average = Double((up + down) / 2)
+    let average = (up + down) / 2.0
     print(average, terminator: " ")
 }
 
 /// 전체 평균
 print()
 
-var total = 0
+var total = 0.0
 
 for number in numbers {
     total += number.reduce(0, + )
